feat: update management page for appointments
